feat: frontend

+++ app/Http/Controllers/ExportController.php
@@ -5,31 +5,37 @@ namespace App\Http\Controllers;
 use App\Models\Export;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Inertia\Inertia;
 
 class ExportController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
-        $exports = Export::paginate(15);
+        $exports = Export::paginate(5);
+
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Export  $export
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Export $export)
+    public function show($export) {
+        $export = Export::find($export);
+
+        return Storage::download($export->file_name);
+    }
+
+    public function destroy($export)
     {
-        Storage::delete($export->file_name);
+        $export = Export::find($export);
 
-        $export->delete();
+        if ($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
 
-        return "deletemo";
+        return redirect()->back()
+            ->with('success', 'Seu arquivo foi removido com sucesso');
+
     }
 }
